refactor : divide the code into functions
