+++ app/svg/stats.ts
